requirements.txt
beautifulsoup4==4.7.1
boto3==1.9.120
botocore==1.12.120
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-constance==2.4.0
django-picklefield==2.0
django-storages==1.7.1
django-summernote==0.8.11.4
docutils==0.14
factory-boy==2.11.1
Faker==1.0.4
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
jmespath==0.9.4
lxml==4.3.2
orcid==1.0.3
Pillow==5.4.1
psycopg2-binary==2.7.7
pyaml==18.11.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
s3transfer==0.2.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.8
text-unidecode==1.2
urllib3==1.24.1
webencodings==0.5.1
